+++ tests/Types/UpdateTest.php
@@ -3,7 +3,6 @@
 namespace TelegramBot\Api\Test\Types;
 
 use TelegramBot\Api\Test\AbstractTypeTest;
-use TelegramBot\Api\Types\BusinessMessagesDeleted;
 use TelegramBot\Api\Test\Types\Inline\ChosenInlineResultTest;
 use TelegramBot\Api\Test\Types\Inline\InlineQueryTest;
 use TelegramBot\Api\Test\Types\Payments\Query\PreCheckoutQueryTest;